+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.31.17
+* **[Fix]** android: autoconnect could enter disconnect / connection loop
+* **[Fix]** android: calling `disconnect` should always disable autoconnect, even if already disconnected
+
+## 1.31.16
+* **[Fix]** forgot to push to mtu stream after disconnection
+
 ## 1.31.15
 * **[Feature]** support advertising `appearance` on Android